requirements.txt
pandas
Pillow>=6.2.0
numpy>=1.8.1,<=1.14.0
scipy>=0.13.3,<=0.17.0
protobuf>=2.5.0,<=3.2.0
six>=1.5.2,<=1.10.0
requests>=2.20.0
gevent>=1.0,<=1.1.0
gevent-websocket>=0.10,<=0.10.1
Flask>=0.12.3
Flask-WTF>=0.11,<=0.12
wtforms>=2.0,<=2.1
Flask-SocketIO==2.6
setuptools>=3.3,<=20.7.0
lmdb==0.87
h5py>=2.2.1,<=2.6.0
psutil>=5.7.0
matplotlib>=1.3.1,<=1.5.2
scikit-fmm>=0.0.9
python-magic>=0.2
boto>=2.48.0
pydot>=1.0.28,<=1.0.29
